FIX: Correct decoherence correction formula
Eq. 17 in the Persico paper is wrong
(https://aip.scitation.org/doi/10.1063/1.2715585)
The factor in eq. 17 should be damping populations, not coefficients
For now, the correct formula is switched on by
'correct_decoherence=.true.' in the sh input

After we test this, we will switch to it permanently,
but should provide a keyword to support reproducing previous
calculations.

Thank you prof. Basile Curchod for pointing this out.