Harden PLUMED restart
Previously, in order to restart the PLUMED simulations,
the 'RESTART' keyword needed to be added at the beginning
of the PLUMED input file, in addition to setting 'irest=1'
in the ABIN input file.

In this diff, we pass the restart info from ABIN to PLUMED directly
so the additional keyword in the PLUMED input is not needed
(but it can still be there).

In addition, I added some basic unit tests for the PLUMED
integration and I improved the current PLUMED test:
1. to test the restart capability
2. print more output files from PLUMED
   to check we're passing correct info from ABIN.
   Specifically, we check masses and energies.

I also added a test for compilation without PLUMED.

Finally, I added the brand new 2.7 PLUMED version to GitHub CI.

Because of these changes, I needed to bump up the
minimum PLUMED API version. Nevertheless, the 2.5 version
still works, and this is the last officially supported version
so I am not worried about compatibility with older versions.

Also, I find out that temperature was passed incorrectly
into PLUMED! (it was set to zero). I commented out the call
so PLUMED should now die if temperature is not specified in
the plumed input. I'll fix this in a future PR. Having
to specify temperature in two different places is suboptimal.