Test TeraChem MPI interface (#62)
Test TC-MPI interface

The primary goal here was to write tests for the TeraChem MPI interface implemented in src/force_tera.F90. There's also a separate interface for Surface Hopping defined in force_terash.F90, but that will be handled in a follow-up PR.

I also ended up doing a little refactor and separated some common functions from force_tera.F90 to tera_mpi_api.F90. The advantage is that they can now be more easily reused in force_terash.F90 and force_tera.F90 and force_terash.F90 are now independent. I also spend quite some time improving the error handling so the interface should now be more robust and easier to debug.

The tricky thing here is of course the TC part. We don't want to run real TC as part of our tests so I essentially ended up extracting the TeraChem side of the API from TeraChem source code and we compile and run it as a fake TC server as part of the tests (see key files below). Moreover, the TC server is compiled together with the q-TIP4P/Fw force field so it returns real forces and energies (of course only for pure water). I reused the same code that is part of ABIN itself and that we use in other tests (pot='mmwater'). Thanks to this trick, I was able to generate the test reference files (*ref) by simply running ABIN with qTIP4P force field, and that way I could verify that the fake TC server works correctly and returns correct energies and forces.

There are several tests:

    tests/TERAPI: The most common use case, classical MD.
    tests/TERAPI-PIMD: Mostly the same as above but with several PI beads. The only difference in terms of the API is that we pass a different scratch directory name for each bead so that each bead has it's separate scratch dir. (the fake TC server does not really populate the scratch directory. Instead, for testing purposes it creates a file with the same name as the scratch directory).
    tests/TERAPI-PIMD-PARALLEL: Same as above, but we connect to several TC servers so that we can run different beads in parallel. This is controlled by the nteraservers variable in the ABIN input.
    tests/TERAPI-FAILS : This is in fact a collection of different tests verifying different failure modes and error handling. Thanks to these tests I was able to achieve almost 100% test coverage (see Codecov report).

There are a lot of new files, but most of them are new test files. In any case, it's probably easier to check out this branch locally instead of browsing it in GitHub UI.

Key files:

    src.force_tera.F90 and src/tera_mpi_api.F90
    tests/tc_mpi_api.cpp, tests/tc_mpi_api.h - TeraChem side of the API
    tests/TERAPI*/tc_server.cpp, the actual code for the server that uses the API and acts as TeraChem (executes in an MD loop).
    each test is driven by it's own test.sh file. Common BASH helper function were extracted to tests/test_tc_server_utils.sh


* Debug option for mpich build
* Simplify handle_mpi_error
* Skip TERAPI test for OpenMPI
* Separate TCServerMock so it can be reused in multiple tests
* Refactor TCServerMock to be more granular
* Hide/remove unused code from force_tera.F90
* Squash compiler warnings for non-MPI compilation from force_terash
* Use common FFLAGS accross jobs in GA
* Print charges and dipole moments in force_tera.F90
Also check for the received count for each MPI_Recv call
in TCServerMock. MPI fails automatically if it receives
more then specified, but allows to receive less than the
buffer capacity so we need to check that manually.
I wonder what else should we be checking.
* Try MPICH 3.4.1
* TCServerMock uses qTIP4PFw for energies and gradients
* Test multiple TC servers with PIMD
There's a bug in hydra_nameserver where
it crashes when multiple TC servers call MPI_Unpublish_name
https://github.com/pmodels/mpich/issues/5058

Hence, we're passing TC port names to ABIN via files.

* GA: Use v3.3.2 and and v3.4.1 MPICH
* Print MPI ERROR string
* Parallelize connection to multiple TS servers via OpenMP
* Add WITHOUT_MPI test
* Do not print dipoles and charges when running multiple TC servers
* none -> _none_
* Remove MAXTERASERVERS restriction
* Increase waiting time
* Print TC and ABIN outputs when failing
* gitignore update
* always restart hydra_nameserver
* Deallocate TC MPI communicators!
* Don't call MPI_Abort with _tera_
* Relaunch nameserver in each subtest
* Log if server name not provided
* Harden MPI Publish + subroutine print_mpi_error
* test mpi_sleep validation
* Add CMDLINE test
* Validate scrdir names
* Check we got all expected data from MPI_Recv
* Test mpich build with ch4:ofi
* Back to ch3:nemesis
* Do not connect to TC servers in parallel
I'm getting segfaults from MPI_Comm_connect,
maybe a bug in MPICH?

On the other hand, we try to parallelize
the TC initialization, sending number of atoms
and atom types at the beginning.

* Remove file interface for TeraChem