Fix GitHub CI
Apparently, GCC-8 has been removed from the
Ubuntu-18.04 base image in favour of GCC-10
so we need to adjust.

We still use GCC-7 since this is what is available
on PHOTOX clusters. That's also why we cannot switch
to Ubuntu-20.04 image as it only provides GCC-9/10

I also needed to tweak the FFLAGS for MPI compilation
with GCC-10. The new GFortran does not like when different
types are passed into the same functions (which is typical
for MPI_Send/MPI_Recv).

Lastly, I tried to fix the misbehaving Zenodo badge in README.