Refactor error handling (#72)
Here we tentatively refactor the error handling in ABIN, with the ultimate goal of replacing abinerror().

The main purpose of this refactor is to allow the unit tests to test for error conditions. This is currently not possible since calling abinerror() unconditionally stops the program, thus disrupting the test suite. To solve that we add a bit of misdirection in src/error.F90. Essentially, we allow pFUnit to swap out our error handler subroutine with its own, which then allows checking assertions in unit tests. This swapping occurs via the set_error_method in unit_tests/throw_with_pfunit.F90.

The new default error handler is called fatal_error() and should be used instead of abinerror. I converted only a couple of instance for now, and if this all work, I will convert the rest in a separate PR.

As an added bonus, the new fatal_error takes different arguments from abinerror. It takes the file name, line number, and error message. The first two arguments can be conveniently provided via preprocessor-defined constants __FILENAME__ and __LINE__. This way, we can provide the exact location in the code where the error occurred (and the cost of slightly more verbose code). That's pretty cool! tada I am open for discussion about how the errors should be phrased exactly so that it's clear. The format for now resembles how e.g. gfortran prints errors.