Fix unitialized variable for inormalmodes==2
I think this mode was broken before this commit,
and perhaps it is still broken, but I am not even sure
what it was supposed to do or if it ever worked. Probably not...

I think it was just testing a normal mode transform with PIMD,
using a slightly different version of the hamiltonian.

Closes #63