Add basic test for TC-MPI surface hopping interface (#81)
I verified that this new test would have caught the bug that was fixed in #80.

As with the previous TC-MPI tests, the new C++ code was copied from TeraChem and trimmed and cleaned up, and I added some basic assertions about what is passed by ABIN. Specificically, in the SH interface, MO coefficients, CI vectors and some other stuff (called 'blob') is passed back and forth so I am checking that the the numbers match. I am also testing the restart capability here. In this interface, there's an extra file wfn.bin which stored the MOs and CI vectors per above.

While the whole interface is tested, I am not actually running non-adiabatic dynamics since that would require implementing some kind of a model for NAC vectors. For this test, we're simply running ground state MD, forces and energies are taken from the in-built water force field, as in previous tests.

TODO: we need to test a proper surface hopping run, maybe by implementing on of the analytical models for NA dynamics, e.g. one of the 1D Tully models, or vibronic coupling model.

I have also added some unit tests, mostly to test various error conditions. I checked on Codecov that most of the code in force_terash.F90 is now tested. There are some lines specific for Landau-Zener, so an additional test will be needed for that in the future.