Fix arrays sizes (nwalk instead of nwalk+1)
Due to historical reasons, we were allocating arrays
with nwalk + 1, where in PIMD transformation routines we
would use the extra slot to ensure the cyclical condition
of the PIMD ring. This made the code confusing and hard to maintain,
so instead I rewrote the routines to encode this condition exclicitly.