Redirect stdout to /dev/null for my_rank>0
I was getting increasingly annoyed about
how we handle output to stdout, especially
when I was writing unit tests and needed to suppress it.
For REMD, we only want the first replica
to write to stdout, so we had a lot of code like:

if (my_rank == 0) write (*, *) 'something'

I found on the internet a brilliant idea - redirect output to /dev/null!
I made a couple of helper functions in files.F90, and now we can do this instead:

 In init.F90
if (my_rank > 0) call stdout_to_devnull()

 Elsewhere
use mod_files, only: stdout
write (stdout, *) 'something'

I also wrote similar helpers for stderr.
Because of this, I was able to get rid of global my_rank entirely,
(in favour of using get_mpi_rank() where needed)

In this commit I converted mostly only write statements
that mattered, so there are still a lot of write (*,*)

Because of various module dependencies, I also needed
move a couple functions into different modules,
but in each of the new placement made more sense.