Use the correct decoherence correction by default
Due to the error in the equation in the original paper,
the decoherence correction was not implemented correctly
(never take anything for granted!).
This was already fixed in commit c7b342b9ad2,
here we switch the use of correct formula by default.

To reproduce old results, add correct_decoherence=.false.
in the input file in namelist sh.

Thanks to Basile Curchod and Antonio Prlj for catching this!