Warning for discontinuities in Landau Zener
LZSH exhibits large problems when discontinuities in energies are
encountered. Simple warning based on large change in second derivatives
is added. It was tested on azobenzene and bilirubin and it was able to
warn when suspicious hops occured. Also warning about too many states
was added at the beginning of the output file.