Bring back TeraChem file interface
The Amber MPI interface in TC was removed in 2021.
We've used this interface for ground state MD.
Instead, we now have the TCPB interface, but still
in an experimental stage. Therefore, for now we bring back
the simple file interface as a quick hassle free way
to run MD with TeraChem.
WARNING: For small / medium sized molecules, you will
lose a lot of time during repeated GPU initialization!