Modifying TERAPI-LZ test
Modification of TERAPI-LZ test to be compatible with the new
modifications of the LZ algorithm. The problem was just three steps.
Thus, the previous second derivative used in the discontinuity check was
still set to 0 which caused the code to think there is a discontinuity.
One more step was artificially added so that it has also the previous
second derivative.