Improvements of LZSH algorithm (#159)
* LZ hopping probability evaluated only to neighbouring states

Current implementation allows for hops like (05-->01), which is not
compatible with the LZ formula that was derived for two-state problem.
This commit changes it and evaluates it only for closest neighbours.

* Warning for discontinuities in Landau Zener

LZSH exhibits large problems when discontinuities in energies are
encountered. Simple warning based on large change in second derivatives
is added. It was tested on azobenzene and bilirubin and it was able to
warn when suspicious hops occurred. Also warning about too many states
was added at the beginning of the output file.